(MP) fontconfig: Use our port's configuration
FC_CACHEDIR and FONTCONFIG_PATH are preprocessed into some of the build
products, although I'm not sure if the static library is one of them.
I think the font paths are only stored in the generated fonts.conf
(which the build discards), but let's maintain consistency with the port
just in case.